requirements.txt
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
click==8.0.1
constantly==15.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-Twisted==0.1.2
greenlet==1.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
mysql-connector-python==8.0.25
observable==1.0.3
packaging==20.9
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pynvim==0.4.3
pyparsing==2.4.7
pytest==6.2.4
six==1.16.0
SQLAlchemy==1.4.19
toml==0.10.2
Twisted==21.2.0
Werkzeug==2.0.1
zope.interface==5.4.0
